refactor: extract gameboard engine and constants
